# assets/openapi.yaml
openapi: 3.0.0
info:
  title: VSCode ChatGPT Plugin
  description: Interact with code in your VSCode Editor
  version: 'v1'
servers:
  - url: 'http://localhost:3000'
paths:
  /tabs:
    get:
      operationId: getAllTabs
      summary: Gets a list of all tabs in the editor
      responses:
        '200':
          description: A list of all tabs and their contents
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    tabName:
                      type: string
                    languageId:
                      type: string
                    isDirty:
                      type: boolean
                    isUntitled:
                      type: boolean
  /tabs/{tabName}:
    get:
      operationId: getTab
      summary: Gets the text of the document related to a specific tab
      parameters:
        - name: tabName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The document text
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Tab not found
  /tabs/modify:
    post:
      operationId: modifyTab
      summary: Patch code on a specific tab 
      requestBody:
        description: The range of text to replace and the new text
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tabName:
                  type: string
                  description: The tab to patch
                startLine:
                  type: integer
                  description: The start line of the range
                startCharacter:
                  type: integer
                  description: The start character of the range
                endLine:
                  type: integer
                  description: The end line of the range
                endCharacter:
                  type: integer
                  description: The end character of the range
                newText:
                  type: string
                  description: The text to replace the range with
      responses:
        '200':
          description: Diff view opened successfully
        '400':
          description: Invalid request body
        '404':
          description: Tab not found
  /tabs/{tabName}/symbols:
    get:
      operationId: getTabSymbols
      summary: Gets the outline of symbols in a document related to a specific tab
      parameters:
        - name: tabName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The document symbols. 
          content:
            text/plain:
              schema:
                type: string
        '404':
          description: Tab not found
  